test: 통합 테스트 리팩토링
