+++ tomcat/src/test/java/nextstep/jwp/presentation/DefaultControllerTest.java
@@ -1,0 +1,41 @@
+package nextstep.jwp.presentation;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+import org.apache.coyote.http11.Http11Processor;
+import org.junit.jupiter.api.Test;
+import support.StubSocket;
+
+class DefaultControllerTest {
+
+    @Test
+    void index() throws IOException {
+        // given
+        final String httpRequest= String.join("\r\n",
+            "GET /index.html HTTP/1.1 ",
+            "Host: localhost:8080 ",
+            "Connection: keep-alive ",
+            "",
+            "");
+
+        final var socket = new StubSocket(httpRequest);
+        final Http11Processor processor = new Http11Processor(socket);
+
+        // when
+        processor.process(socket);
+
+        // then
+        final URL resource = getClass().getClassLoader().getResource("static/index.html");
+        var expected = "HTTP/1.1 200 OK \r\n" +
+            "Content-Type: text/html;charset=utf-8 \r\n" +
+            "Content-Length: 5564 \r\n" +
+            "\r\n"+
+            new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
+
+        assertThat(socket.output()).isEqualTo(expected);
+    }
+}
